requirements.txt
flask
spacy
sentence-transformers
scikit-learn
python-docx
reportlab
numpy